修复数据库 migration 的数据缺失, 专业表

+++ database/migrations/2019_10_30_164213_create_school_configurations_table.php
@@ -54,9 +54,11 @@ class CreateSchoolConfigurationsTable extends Migration
         Schema::dropIfExists('school_configurations');
 
         Schema::table('departments', function (Blueprint $table) {
-            $table->dropColumn(ConfigurationTool::KEY_OPTIONAL_COURSES_PER_YEAR);
-            $table->dropColumn(ConfigurationTool::KEY_SELF_STUDY_NEED_REGISTRATION);
-            $table->dropColumn(ConfigurationTool::KEY_STUDY_WEEKS_PER_TERM);
+            if(Schema::hasColumn('departments',ConfigurationTool::KEY_OPTIONAL_COURSES_PER_YEAR)){
+                $table->dropColumn(ConfigurationTool::KEY_OPTIONAL_COURSES_PER_YEAR);
+                $table->dropColumn(ConfigurationTool::KEY_SELF_STUDY_NEED_REGISTRATION);
+                $table->dropColumn(ConfigurationTool::KEY_STUDY_WEEKS_PER_TERM);
+            }
         });
     }
 }
